+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.80.0
+
+* Reinforce error handling and reduce usage of panic in codegen #1298 (thanks @Desdaemon)
+* fix: sync return with Freezed metadata #1304 (thanks @SiongSng)
+
 ## 1.79.0
 
 * Update example tab on pub.dev and remove low-quality dependency #1263